requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl_cffi==0.13.0
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
dotenv==0.9.9
frozendict==2.4.7
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
loguru==0.7.3
logzero==1.7.0
multitasking==0.0.12
numpy==2.3.5
openai==2.9.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyotp==2.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
smartapi-python==1.5.5
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
win32_setctime==1.2.0
yfinance==0.2.66
